fix: 修复编译问题，更新span数据类型

+++ 2015_test/Program.cs
@@ -47,7 +47,7 @@ namespace ConsoleProxy
         public bool trade;
         public int closevolumn;
         public int openvolumn;
-        public int span;
+        public double span;
 
     }
